feat: Add listLocalPeers() method
I realized we do not have a way to list connected peers, which is
necessary to implement the invite screen. Initial proposed
implementation is to return a list of all connected peers: it will be
the responsibility of the client to list project members and filter this
list to not display connected devices that are members of the project.